ProjectAccessor accepts repositoryAccessors as parameter in methods
ProjectAccessor is not tied with one repository anymore, so it is not feasible for it to have just on repositoryAccessor as a field. Instead, the repositoryAccessor is accepted as a parameter, which leads to greater abstraction, as the projectAccessor is not aware (and doesnt need to be) of the issue and repository type it is working with.