Multiple Trello Columns
When connecting new repository to a project, you have the option to pass a list of parameters and map the columns names from the service
to the states used in tasksyncer and displayed at trello.
Working with all four connected services - Gitlab, Github, Jira, Trello
With Gitlab and github is also checked, if the provided mapping is valid.